refactor dpm mandate routes
